+++ src/sdk/PnP.Core.Admin/Model/SharePoint/Core/Public/ISharePointAdmin.cs
@@ -111,32 +111,46 @@ namespace PnP.Core.Admin.Model.SharePoint
         #region Get Search Configuration
 
         /// <summary>
-        /// Gets the search configuration of the Tenant
+        /// Gets the search configuration of the tenant
         /// </summary>
         /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
         /// <returns>Search configuration XML</returns>
         Task<string> GetTenantSearchConfigurationXmlAsync(VanityUrlOptions vanityUrlOptions = null);
 
         /// <summary>
-        /// Gets the search configuration of the Tenant
+        /// Gets the search configuration of the tenant
         /// </summary>
         /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
         /// <returns>Search configuration XML</returns>
         string GetTenantSearchConfigurationXml(VanityUrlOptions vanityUrlOptions = null);
 
         /// <summary>
-        /// Gets the managed properties from the search configuration of the Tenant
+        /// Gets the managed properties from the search configuration of the tenant
         /// </summary>
         /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
         /// <returns>List of managed properties</returns>
         Task<List<IManagedProperty>> GetTenantSearchConfigurationManagedPropertiesAsync(VanityUrlOptions vanityUrlOptions = null);
 
         /// <summary>
-        /// Gets the managed properties from the search configuration of the Tenant
+        /// Gets the managed properties from the search configuration of the tenant
         /// </summary>
         /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
         /// <returns>List of managed properties</returns>
         List<IManagedProperty> GetTenantSearchConfigurationManagedProperties(VanityUrlOptions vanityUrlOptions = null);
+
+        /// <summary>
+        /// Sets the search configuration for the tenant
+        /// </summary>
+        /// <param name="configuration">Search configuration, obtained via <see cref="GetTenantSearchConfigurationXml"/> to apply</param>
+        /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
+        Task SetTenantSearchConfigurationXmlAsync(string configuration, VanityUrlOptions vanityUrlOptions = null);
+
+        /// <summary>
+        /// Sets the search configuration for the tenant
+        /// </summary>
+        /// <param name="configuration">Search configuration, obtained via <see cref="GetTenantSearchConfigurationXml"/> to apply</param>
+        /// <param name="vanityUrlOptions">Optionally specify the custom vanity URI's used by this tenant</param>
+        void SetTenantSearchConfigurationXml(string configuration, VanityUrlOptions vanityUrlOptions = null);
 
         #endregion