+++ src/sdk/PnP.Core.Test/SharePoint/SiteTests.cs
@@ -981,6 +981,21 @@ namespace PnP.Core.Test.SharePoint
                 Assert.IsNotNull(mps);
             }
         }
+
+        [TestMethod]
+        public async Task SetWebSearchConfigurationTest()
+        {
+            //TestCommon.Instance.Mocking = false;
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite))
+            {
+                var searchConfigXml = context.Site.GetSearchConfigurationXml();
+
+                Assert.IsNotNull(searchConfigXml);
+                Assert.IsTrue(!string.IsNullOrEmpty(searchConfigXml));
+
+                context.Site.SetSearchConfigurationXml(searchConfigXml);
+            }
+        }
         #endregion
     }
 }