+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IWeb.cs
@@ -1316,6 +1316,17 @@ namespace PnP.Core.Model.SharePoint
         /// <returns>List of managed properties</returns>
         List<IManagedProperty> GetSearchConfigurationManagedProperties();
 
+        /// <summary>
+        /// Sets the search configuration for the web
+        /// </summary>
+        /// <param name="configuration">Search configuration, obtained via <see cref="GetSearchConfigurationXml"/> to apply</param>
+        Task SetSearchConfigurationXmlAsync(string configuration);
+
+        /// <summary>
+        /// Sets the search configuration for the web
+        /// </summary>
+        /// <param name="configuration">Search configuration, obtained via <see cref="GetSearchConfigurationXml"/> to apply</param>
+        void SetSearchConfigurationXml(string configuration);
         #endregion
 
         #endregion