+++ src/sdk/CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Option to convert an `IFile` into another format (PDF, JPG, Html or GLB) and process the returned stream (e.g. download/upload as file) #849 [MathijsVerbeeck - Mathijs Verbeeck]
 - Get preview from `IFile` for embedding #885 [MathijsVerbeeck - Mathijs Verbeeck]
 - Support for querying the `IWeb` recycle bin #894 [jansenbe - Bert Jansen]
+- Support for exporting and importing the search configuration at web, site and tenant level #887 [jansenbe - Bert Jansen]
 
 ### Changed